Multi_DB_Router/requirements.txt
Django==1.10
django-multiselectfield==0.1.12
mysql-connector-python==8.0.6
mysqlclient==1.4.6
psycopg2==2.7.4
PyMySQL==0.8.0
python-dotenv==0.13.0
